requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
biopython==1.84
boltz @ git+https://github.com/jwohlwend/boltz.git@cb04aeccdd480fd4db707f0bbafde538397fa2ac
certifi==2026.1.4
charset-normalizer==3.4.4
chembl-structure-pipeline==1.2.2
click==8.1.7
cuequivariance==0.8.0
cuequivariance-ops-cu12==0.8.0
cuequivariance-ops-torch-cu12==0.8.0
cuequivariance-torch==0.8.0
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.8.0
einx==0.3.0
exceptiongroup==1.3.1
fairscale==0.4.13
fastapi==0.128.0
filelock==3.20.2
freesasa==2.2.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
gemmi==0.6.5
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hydra-core==1.3.2
idna==3.11
ihm==2.8
Jinja2==3.1.6
joblib==1.5.3
lightning-utilities==0.15.2
llvmlite==0.44.0
MarkupSafe==3.0.3
mashumaro==3.14
modelcif==1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==5.29.5
psutil==7.2.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0
pytz==2025.2
PyYAML==6.0.2
rdkit==2025.9.3
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.48.0
setproctitle==1.3.7
six==1.17.0
smmap==5.0.2
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchmetrics==1.8.2
tqdm==4.67.1
triton==3.5.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
wandb==0.18.7
yarl==1.22.0
